Initial repository restructuring
- Moved contents out of the root src/ directory.
- Removed previous go module config files.
- Removed Makefile for submitting labs for inspection

This repository is primarily meant for self study. Scripts and artefacts
related to official MIT lab submission and course grading process have
been removed since they don't have any use in the given context. However,
all the scripts for testing the implementation will be retained. The
grades produced by these scripts are purely for the sake of verifying
the level of correctness of the implementation.